+++ CHANGELOG.md
@@ -8,3 +8,6 @@ TBD
 - Fix: Update all pointer uses in ast_tree to avoid memory leaks
 - Fix: Update all pointers around the LSP to avoid memory leaks
 - Added: Include support for pass params validation
+- Added: Include support for material params validation
+- Added: Include support for technique params validation
+- Fix: Issue with allow match_literals on params list